Refactor code and add new components
