+++ app/(site)/page.tsx
@@ -1,5 +1,5 @@
 import { getProfile } from "@/sanity/sanity.fetch";
-import HeroSvg from "../../../components/icons/HeroSvg";
+import HeroSvg from "../../components/icons/HeroSvg";
 import TransitionEffect from "@/components/TransitionEffect";
 import AnimatedText from "@/components/AnimatedText";
 import HireMe from "@/components/shared/HireMe";
